feat: rename web captions to caption generator
