+++ mentors/src/data_pipeline/caption_generator.py
